refactor(post): restructure DTOs and entities for car and bike features
- Moved car and bike DTOs to dedicated directories under `post-details`.
- Renamed `color_id` to `color` across DTOs, mapper, and entities.
- Removed `trim_id` field from car and bike DTOs and entities.
- Added new `SEAT_OPTIONS` constant for seat validation and updated related logic.
- Improved TypeScript typings and added `Transform` utilities for `seats` handling.
