+++ apps/api/src/modules/post-details/dto/battery/battery-detail-response.dto.ts
