+++ apps/api/src/modules/post-details/dto/bike/bike-details-response.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { BodyStyle, Origin, VehicleColor } from '../../../../shared/enums/vehicle.enum';
+import { BikeStyle, Origin, VehicleColor } from '../../../../shared/enums/vehicle.enum';
 
-export class CarDetailsResponseDto {
+export class BikeDetailsResponseDto {
   @ApiProperty({ nullable: true })
   brand_id!: number | null;
 
@@ -9,22 +9,16 @@ export class CarDetailsResponseDto {
   model_id!: number | null;
 
   @ApiProperty({ nullable: true })
-  trim_id!: number | null;
-
-  @ApiProperty({ nullable: true })
   manufacture_year!: number | null;
 
-  @ApiProperty({ enum: BodyStyle, nullable: true })
-  body_style!: BodyStyle | null;
+  @ApiProperty({ enum: BikeStyle, nullable: true })
+  bike_style!: BikeStyle | null;
 
   @ApiProperty({ enum: Origin, nullable: true })
   origin!: Origin | null;
 
   @ApiProperty({ enum: VehicleColor, nullable: true })
-  color_id!: VehicleColor | null;
-
-  @ApiProperty({ nullable: true })
-  seats!: number | null;
+  color!: VehicleColor | null;
 
   @ApiProperty({ nullable: true })
   license_plate!: string | null;
@@ -42,10 +36,10 @@ export class CarDetailsResponseDto {
   range_km!: number | null;
 
   @ApiProperty({ nullable: true })
-  charge_ac_kw!: number | null;
+  motor_power_kw!: number | null;
 
   @ApiProperty({ nullable: true })
-  charge_dc_kw!: number | null;
+  charge_ac_kw!: number | null;
 
   @ApiProperty({ nullable: true })
   battery_health_pct!: number | null;
