+++ apps/api/src/modules/posts/dto/bike/create-post-bike.dto.ts
@@ -2,7 +2,7 @@ import { ValidateNested } from 'class-validator';
 import { Type } from 'class-transformer';
 import { BaseCreatePostDto } from '../base-create-post.dto';
 import { PostType } from '../../../../shared/enums/post.enum';
-import { CreateBikeDetailsDto } from '../../../post-details/dto/create-bike-details.dto';
+import { CreateBikeDetailsDto } from '../../../post-details/dto/bike/create-bike-details.dto';
 
 export class CreateBikePostDto extends BaseCreatePostDto {
   postType: PostType = PostType.EV_BIKE;
