+++ apps/api/src/modules/posts/dto/car/create-post-car.dto.ts
@@ -2,7 +2,7 @@ import { ValidateNested } from 'class-validator';
 import { Type } from 'class-transformer';
 import { BaseCreatePostDto } from '../base-create-post.dto';
 import { PostType } from '../../../../shared/enums/post.enum';
-import { CreateCarDetailsDto } from '../../../post-details/dto/create-car-details.dto';
+import { CreateCarDetailsDto } from '../../../post-details/dto/car/create-car-details.dto';
 
 export class CreateCarPostDto extends BaseCreatePostDto {
   postType: PostType = PostType.EV_CAR;
