+++ apps/api/src/shared/constants/vehicle.ts
@@ -1,0 +1,3 @@
+// shared/constants/vehicle.ts
+export const SEAT_OPTIONS = [2, 4, 5, 6, 7, 8, 9, 10, 12, 14, 16] as const;
+export type SeatOption = (typeof SEAT_OPTIONS)[number];
